Create migration and model Category

+++ database/migrations/2019_05_26_192052_create_roles_user_table.php
@@ -14,8 +14,8 @@ class CreateRolesUserTable extends Migration
     public function up()
     {
         Schema::create('roles_user', function (Blueprint $table) {
-            $table->unsignedInteger('userId');
-            $table->unsignedInteger('roleId');
+            $table->bigInteger('userId');
+            $table->bigInteger('roleId');
             $table->unique(['userId', 'roleId']);
             $table->foreign('userId')->references('id')->on('users');
             $table->foreign('roleId')->references('id')->on('roles');
